+++ CHANGELOG.md
@@ -8,6 +8,17 @@ the following form: YYYY.0M.0D.
 
 ## [Unreleased]
 
+## [2025.04.16]
+
+### Added
+
+- Designed test suite for MChest component
+- Designed two different use cases for MChest component
+
+### Updated
+
+- Changed design to handle errors in both `addItem()` and `removeItem()` methods
+
 ## [2025.04.09]
 
 ### Added
@@ -51,6 +62,7 @@ and `hashCode()`
 - Designed a Stock Trader component
 - Designed a Playing Card component
 - Designed a Minecraft Chest component
+
 ## [2024.12.30]
 
 - Added table-based rubrics to all 6 parts of the project